==> 2-descriptive-outcomes/wasting/0-run-descriptive-outcomes-wasting.R <==

rm(list=ls())
source(paste0(here::here(), "/0-config.R"))


source(here("2-descriptive-outcomes/wasting/1_wast_calcinc.R"))
source(here("2-descriptive-outcomes/wasting/2_birth_wast_outcomes.R"))
source(here("2-descriptive-outcomes/wasting/3_wast_summarize_outcomes.R"))
source(here("2-descriptive-outcomes/wasting/4_co_summarize_outcomes.R"))
source(here("2-descriptive-outcomes/wasting/5_co_flowdata.R"))
source(here("2-descriptive-outcomes/wasting/7_wast_ir_sensitivity.R"))
source(here("2-descriptive-outcomes/wasting/8_wast_manuscript_extra_stats.R"))
source(here("2-descriptive-outcomes/wasting/9_create_wast_seasonality_datasets.R"))
source(here("2-descriptive-outcomes/wasting/10_wast_summarize_outcomes_FE.R"))
source(here("2-descriptive-outcomes/wasting/12_co_summarize_outcomes_FE.R"))
source(here("2-descriptive-outcomes/wasting/13_create_birth_wast_seasonality_dataset.R"))
source(here("2-descriptive-outcomes/wasting/13_birth_Zscore_diff_by_month.R"))
source(here("2-descriptive-outcomes/wasting/21_uwt_summarize_outcomes.R"))



# #Check:
# source(here("2-descriptive-outcomes/wasting/8_wast_manuscript_extra_stats.R"))
# source(here("2-descriptive-outcomes/wasting/9_create_wast_seasonality_datasets.R"))

#supplimentary
 source(here("2-descriptive-outcomes/wasting/supplimentary analyses/1_wast_summarize_outcomes_no_Kenaba_bw.R"))
 source(here("2-descriptive-outcomes/wasting/supplimentary analyses/2_co_summarize_outcomes_no_Kenaba.R"))
 source(here("2-descriptive-outcomes/wasting/supplimentary analyses/3_sex_strat_wast_summarize_outcomes.R"))
